# VideoGameExchange/VideoGameExchange.Server/Models/Link.cs
namespace VideoGameExchange.Server.Models;

public class Link
{
    public string Href { get; set; } = string.Empty;
    public string Rel { get; set; } = string.Empty;
    public string Method { get; set; } = "GET";
}