# VideoGameExchange/VideoGameExchange.Server/Models/PatchGame.cs
namespace VideoGameExchange.Server.Models;

public class PatchGame
{
    public string? Title { get; set; }
    public string? Platform { get; set; }
    public string? Condition { get; set; }
    public int? Year { get; set; }
    public string? Publisher { get; set; }
}