# VideoGameExchange/VideoGameExchange.Server/Models/Problem.cs
namespace VideoGameExchange.Server.Models;

public class Problem
{
    public string? Type { get; set; }
    public string? Title { get; set; }
    public int? Status { get; set; }
    public string? Detail { get; set; }
    public string? Instance { get; set; }
}